\name{accelmissing-package}
\alias{accelmissing-package}
\alias{accelmissing}
\docType{package}
\title{
Missing Value Imputation for Accelerometer Data
}
\description{
We present a statistical method for imputing missing values in accelerometer data. The methodology includes both parametric and semi-parametric multiple imputation under the zero-inflated Poisson lognormal model. It also offers several functions to preprocess accelerometer data before imputation. These include detecting wear and non-wear time, selecting valid days and subjects, and generating plots.
}
\details{
\tabular{ll}{
Package: \tab accelmissing\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2025-05-30 \cr
License: \tab GPL (>=2)\cr
}
}
\author{
Jung Ae Lee <jungae.lee@umassmed.edu>

Maintainer: Jung Ae Lee <jungaeleeb@gmail.com>
}
\references{
Lee JA, Gill J (2016). Missing value imputation for physical activity data measured by accelerometer. Statistical Methods and Medical Research.
}
\keyword{ accelerometer }
\keyword{ physical activity }
\keyword{ missing count data }
\seealso{
\code{mice}, \code{pscl}
}
