\name{accelmissing-package}
\alias{accelmissing-package}
\alias{accelmissing}
\docType{package}
\title{
accelmissing
}
\description{
This packages provides a statistical method to impute the missing values in accelerometer data. The methodology includes both parametric and semi-parametric multiple imputations under the zero-inflated poisson lognormal model. It also provides multiple functions to preprocess the accelerometer data previous to the missing data imputation. These includes detecting wearing and nonwearing time, selecting valid days and subjects, and creating plots.
}
\details{
\tabular{ll}{
Package: \tab accelmissing\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-07-05 \cr
License: \tab GPL (>=2)\cr
}
}
\author{
Jung Ae Lee <jungaeleeb@gmail.com>
}
\references{
Lee JA, Gill J (2016). Missing value imputation for physical activity data measured by accelerometer. Statistical Methods and Medical Research.
}
\keyword{ accelerometer }
\keyword{ physical activity }
\keyword{ missing count data }
\seealso{
\code{mice}, \code{pscl}
}
